+++ src/Player.h
@@ -1,6 +1,8 @@
 #ifndef _PLAYER_H_
 #define _PLAYER_H_
 
+#include <stdint.h>
+
 typedef struct {
     double x;
     double y;
@@ -9,6 +11,6 @@ typedef struct {
 } Player;
 
 void init_player(Player *player, double x, double y);
-void update_player(Player *player, int input_x_direction, int input_y_direction);
+void update_player(Player *player, int8_t input_x_direction, int8_t input_y_direction);
 
 #endif // _PLAYER_H_
